+++ app/(tabs)/List.js
@@ -5,9 +5,6 @@ import {
   View,
   Text,
   Image,
-  Modal,
-  Pressable,
-  TextInput,
 } from "react-native";
 import MainTitle from "@/components/MainTitle";
 import { listImage } from "@/assets/images/list";
@@ -15,6 +12,7 @@ import { useUser } from "@/contexts/UserContext";
 import { useEffect, useState } from "react";
 import {
   AddStudentBtn,
+  AddStudentModal,
   ListEltForStudent,
   ListEltForStudentAccept,
   ListEltForTeacher,
@@ -71,7 +69,6 @@ export default function List() {
   const showRole = userRole == "학생" ? "선생님" : "학생";
 
   const [modalVisible, setModalVisible] = useState(false);
-  const [studentId, setStudentId] = useState("");
 
   const list = data.filter((elt) => elt.connectionStatus === "ACCEPTED");
   const waitingList = data.filter(
@@ -87,47 +84,9 @@ export default function List() {
   // 모달 여닫기 함수
   const toggleModal = () => setModalVisible(!modalVisible);
 
-  // 학생 추가하는 함수
-  const addStudent = () => {
-    // 오류1 : 이미 추가한 사용자인 경우
-    // 오류2 : 없는 사용자인 경우
-    // 성공 : 학생 추가 신청 보내기
-    toggleModal();
-  };
-
   return (
     <SafeAreaView style={{ flex: 1, backgroundColor: "white" }}>
-      {/* 학생 추가 모달창 */}
-      {userRole == "선생님" && (
-        <Modal visible={modalVisible} transparent animationType="fade">
-          <Pressable style={styles.modalBackground} onPress={toggleModal}>
-            <Pressable
-              style={styles.modalContainer}
-              onPress={(e) => e.stopPropagation()}
-            >
-              <Text style={{ fontSize: 16 }}>학생 ID를 입력해주세요</Text>
-              <View style={styles.idInputBox}>
-                <Image
-                  source={listImage.studentIdIcon}
-                  style={{ width: 16, height: 16 }}
-                />
-                <TextInput
-                  placeholder="학생 ID"
-                  value={studentId}
-                  onChangeText={setStudentId}
-                  style={{ fontSize: 16 }}
-                  autoCapitalize="none"
-                  autoCorrect={false}
-                  maxLength={20}
-                />
-              </View>
-              <Pressable onPress={addStudent} style={styles.confirmBtn}>
-                <Text style={{ color: "white" }}>확인</Text>
-              </Pressable>
-            </Pressable>
-          </Pressable>
-        </Modal>
-      )}
+      {modalVisible && <AddStudentModal toggleModal={toggleModal} />}
 
       {/* List 탭 본문 */}
       <ScrollView style={styles.container}>
@@ -196,37 +155,6 @@ const styles = StyleSheet.create({
     paddingHorizontal: 27,
     paddingBottom: 32,
     backgroundColor: "white",
-  },
-
-  modalBackground: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-    backgroundColor: "rgba(0,0,0,0.3)",
-  },
-  modalContainer: {
-    backgroundColor: "white",
-    borderRadius: 10,
-    padding: 24,
-    width: 280,
-  },
-  idInputBox: {
-    width: "100%",
-    flexDirection: "row",
-    gap: 8.6,
-    alignItems: "center",
-    backgroundColor: "#F2F2F2",
-    paddingLeft: 11,
-    paddingVertical: 8,
-    marginVertical: 10,
-  },
-  confirmBtn: {
-    width: "100%",
-    height: 34,
-    alignSelf: "center",
-    justifyContent: "center",
-    alignItems: "center",
-    backgroundColor: "black",
   },
 
   mainTitleContainer: {
