+++ components/ListTab/index.js
@@ -1,3 +1,4 @@
 export * from "./ListElt";
 export * from "./NoList";
 export * from "./AddStudentBtn";
+export * from "./AddStudentModal";
